requirements.txt
boto3==1.26.65
fastapi==0.94.1
pydantic==1.10.5
python-dotenv==1.0.0
uvicorn==0.21.1